some minor tweaks
git-svn-id: https://svn.code.sf.net/p/vpinball/code/trunk@1203 14438d79-0ffd-4333-8ae7-ea2bdd901387

+++ physconst.h
@@ -4,10 +4,10 @@
 //#define PHYSICS_STEPTIME    10000	// usecs to go between each physics update
 #define PHYSICS_STEPTIME    1000	// usecs to go between each physics update
 
-#define PHYSICS_STEPTIME_S  ((float)(PHYSICS_STEPTIME * 1e-6))     // step time in seconds
+#define PHYSICS_STEPTIME_S  (PHYSICS_STEPTIME * 1e-6)     // step time in seconds
 
 #define DEFAULT_STEPTIME      10000      // default physics rate: 100Hz
-#define DEFAULT_STEPTIME_S    0.01f      // default physics rate: 100Hz
+#define DEFAULT_STEPTIME_S    0.01       // default physics rate: 100Hz
 
 #define PHYS_FACTOR         (PHYSICS_STEPTIME_S / DEFAULT_STEPTIME_S)
